Transformer - implement Unitransformer.inputs_full_attention.  This is an option for seq-to-seq as a language model.  Each example         consists of [<inputs>, EOS=1, <targets>, EOS=1].  In the self-attention         layers, positions in the inputs portion of the sequence can see the         entire inputs portion, while positions in the targets portion of the         sequence cannot see future positions.
To do this, we need to decouple the attention visibility logic from the position in the sequence.  We add tensors "read_priority" and "write_priority".
       in self-attention, position a can see position b iff
        read_priority[a] >= write_priority[b]
For fully-autoregressive language models, we set the priorities equal to the position.  For input-full-attention, we set the priority to 0 for the input portion of the sequence and to position for the target portion.

Also add an option Unitransformer.loss_on_targets_only to compute losses only on the <targets> portion of a language model, as defined above.  This is orthogonal to the inputs_full_attention.

PiperOrigin-RevId: 261753655