This CL implements Funnel Transformer introduced in https://arxiv.org/abs/2006.03236.
The main idea is to reduce the sequence dimension to reduce the FLOPS in the
encoder. Optionally, these saved compute can be reinvested in the additional
encoder layers.

All the Funnel-Transformer-specific logics are included in the
FunnelTransformerLayerStack.

There are some subtle differences in this implementation and the paper.

1) In the original paper, all the experiments were run with the encoder-only
model. In this CL, we use it to is to compress the encoder while the decoder
stays the same.

2) A simplified relative attention introduced in Raffel et al.
(https://arxiv.org/abs/1910.10683) is used.

3) Pooling is applied after the self attention.

See the class docstring of FunnerTransformerLayerStack for more details.

PiperOrigin-RevId: 339076573