Implement state update mechanisms for convolution-based incremental decoding.
For incremental decoding, conv layers in the decoder have to record the
recurrent states to be used by the same layer in the next time step. Although
there are multiple variants of conv layer, this logic does not depend on the
specifics of the conv layer. The only relevant information is the filter size
of the layer, which determines how far back the recurrent states should be.

This means that instead of each conv layer having its own logic for incremental
decoding, we can have a common class that implements storing and updating the
state information for all conv layers. This CL implements such class, `Conv1D`.
A conv layer can inherit Conv1D and call `record_states_first_part_mode` during
the first part mode and `record_states_incremental_mode' during the incremental
mode.

PiperOrigin-RevId: 341445785