Implement incremental decoding for various convolutional layers.
This CL implements the incremental decoding for Conv1DLayer and
SeparableConv1DLayer, which is used by LightweightConv1DLocalAttn and
DynamicConv1DLocalAttn.

During decoding, the input `x` does not have the sequence dimension because it
correspond to the current decoding step. This means that we the convolution
operation can be simplified with "VALID" padding. Conv1DLayer supports this
whereas SeparableConv1DLayer does not because the current implementation of the
depthwise convolution does not allow "VALID" padding.

PiperOrigin-RevId: 341449054